% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finance}
\alias{finance}
\title{Pseudo activity count (per minute) data for cats}
\format{A data frame with 756 rows and 50 variables:}
\usage{
finance
}
\description{
A dataset containing the daily closing price of 50 NYSE stocks
from 2014-01-10 to 2017-01-01
}
\keyword{datasets}
