% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvdnorm}
\alias{mvdnorm}
\title{multivariate normal density}
\usage{
mvdnorm(x, mean, sigma, logd)
}
\arguments{
\item{x}{matrix with each row as an observed vector of multivariate normal RVs}

\item{mean}{vector of means}

\item{sigma}{variance-covariance matrix}

\item{logd}{whether log transformation should be used}
}
\value{
a vector of density values for each observed vector
}
\description{
multivariate normal density
}
