% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{package-rarhsmm}
\alias{package-rarhsmm}
\alias{rarhsmm-package}
\alias{package-rarhsmm-package}
\title{Regularized Autoregressive Hidden Semi Markov Models}
\description{
\tabular{ll}{
Package: \tab rarhsmm \cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2017-04-20\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\author{
Zekun Xu \email{zekunxu@gmail.com}

Ye Liu \email{yliu87@ncsu.edu}
Maintainer: Zekun Xu \email{zekunxu@gmail.com}
}
\keyword{Markov}
\keyword{autoregressive,}
\keyword{hidden}
\keyword{models,}
\keyword{regularized}
\keyword{semi-Markov}
