% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmultinomial}
\alias{rmultinomial}
\title{multinomial random variable generator}
\usage{
rmultinomial(n, k, prob, label)
}
\arguments{
\item{n}{number of random variables to generate}

\item{k}{number of categories}

\item{prob}{vector of probabilities summing up to 1}

\item{label}{vector of labels for each category}
}
\value{
multinomial random variables
}
\description{
multinomial random variable generator
}
