% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finance}
\alias{finance}
\title{NYSE stock closing price data}
\format{A data frame with 504 rows and 51 variables:}
\usage{
finance
}
\description{
A dataset containing the daily closing price of 50 NYSE stocks
from 2015-01-02 to 2016-12-30. The first 50 columns are the daily closing
price for each of the 50 stocks, and the last column is the date.
}
\details{
The abbreviation for the 50 stocks are as follows:
"AMD","BAC","AAPL","FTR","BBRY","GE","RAD","MU","CHK","F",
"VALE","FCX","PBR","XOM","INTC","MSFT","WWAV","QQQ","ABEV",
"VZ","HPQ","KKFC","PFE","SWN","T","AKS","FCAU","JPM","SFUN",
"LULU","MT","WFT","CLF","SNA","S","C","HCP","DRYS","FMC",
"CSCO","KMI","AES","X","SIRI","WLL","COP","BSX","JCP","RF","EDUC"
}
\keyword{datasets}
